requirements.txt
# AdvisorIQ — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core ──────────────────────────────────────────────────────────────────────
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0

# ── ML ────────────────────────────────────────────────────────────────────────
scikit-learn>=1.3.0

# ── Database ──────────────────────────────────────────────────────────────────
bcrypt>=4.0.0
psycopg2-binary>=2.9.9     # PostgreSQL (comment out if using SQLite only)

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0

# ── File export ───────────────────────────────────────────────────────────────
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# ── Google Sheets ─────────────────────────────────────────────────────────────
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0

# ── WhatsApp ──────────────────────────────────────────────────────────────────
twilio>=8.10.0

# ── Background Jobs ───────────────────────────────────────────────────────────
apscheduler>=3.10.0
pytz>=2024.1

# ── Payments ──────────────────────────────────────────────────────────────────
razorpay>=1.4.0

# ── Utils ─────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0       # Load .env file
requests>=2.31.0
